YLMF-PC Bot Protection
This script can be applied to block the bot known as PushDo and Cutwail sometimes. YLMF-PC is a fingerprint from this spamming bots and this script has been made to block it on servers that uses CPANEL, if you want to use this script on other servers you should check the directory that stores ACL files and do it manualy, or create a new script to share with us :).